move helpers.js to example/lib/helpers.js (tidy up)
