+++ example/lib/helpers.js
