+++ example/server.js
@@ -1,6 +1,6 @@
 var port = process.env.PORT || 1337; // let heroku define port or use 1337
-var http = require('http');      // core node.js http (no frameworks)
-var app  = require('./helpers'); // auth, token verification & render helpers
+var http = require('http');          // core node.js http (no frameworks)
+var app  = require('./lib/helpers'); // auth, token verification & render helpers
 
 module.exports = http.createServer(function (req, res) {
   var url = req.url;
